% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clean-bib.R
\name{clean_bib}
\alias{clean_bib}
\title{Clean the BibTex exported from mendeley}
\usage{
clean_bib(file_name)
}
\arguments{
\item{file_name}{characters, the name of the bib file}
}
\description{
Clean the BibTex exported from mendeley
}

