% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list2df.R
\name{list2df}
\alias{list2df}
\title{Convert a list to a data frame.}
\usage{
list2df(l, use.names = fill, fill = FALSE, idcol = NULL)
}
\arguments{
\item{l}{A list constaining data.table, data.frame or list object or vector.}

\item{use.names}{If TRUE items will be bound by matching column names.}

\item{fill}{If TRUE fills missing columns with NAs.}

\item{idcol}{Generates an index column. Default (NULL) is not to.}
}
\value{
A data frame.
}
\description{
Convert a list to a data frame. Just a wrapper of rbindlist in data.table
}
\examples{
df <- list2df(list(w1 = list(a = 1, b = 2, c = 3), w2 = list(a = 4, b = 5, c = 6)))
}

