% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/header.R
\name{make_header}
\alias{make_header}
\title{Automatically number headers in markdown file}
\usage{
make_header(n = 1)
}
\arguments{
\item{n}{The number of the first header 3.Default 1.}
}
\value{
A function with a integer parameter(3, 4, 5) that can automatically number headers
}
\description{
Returns a function that can automatically number headers. Only support header 3, 4, 5.
}
\examples{
h <- make_header()
# write in rmarkdown
`r h(3)` H3

`r h(4)` H4
}

