% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outerize.R
\name{outerize}
\alias{outerize}
\title{Outerize a Scalar Function}
\usage{
outerize(FUN, outerize.args)
}
\arguments{
\item{FUN}{function to apply, found via \code{\link[base]{match.fun}}}

\item{outerize.args}{A character vector of arguments which should be outerized.}
}
\value{
A function with the same arguments as \code{FUN}.
}
\description{
Create a function wrapper that outerizes the action of its argument \code{FUN}.
}
\details{
\code{outerize} is just function that calls \code{\link[base]{Vectorize}} recursively. 
Like \code{\link{Vectorize}}, it can not be used with primitive functions and 
the functions that have arguments named \code{FUN}, \code{vectorize.args}, 
\code{SIMPLIFY}, \code{USE.NAMES}.
}
\examples{
f <- function(x, y) x + y
f_outerized <- outerize(f, c("x", "y"))
f_outerized(c(1, 2), c(3, 4, 5))
}

