% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollapply.R
\name{roll_apply}
\alias{roll_apply}
\title{Rolling apply a function on a data frame or matrix.}
\usage{
roll_apply(data, width, .fun, ..., parallel = FALSE, .export = NULL,
  .packages = NULL, progress = FALSE)
}
\arguments{
\item{data}{A data frame or a matrix.}

\item{width}{The rolling window size.}

\item{.fun}{function whose first parameter is the \code{data}.}

\item{...}{Additional parameter passed to \code{.fun}}

\item{parallel}{parallel computing. Default is FALSE. If parallel is positive, 
parallel is the number of cores. If parallel is negative, the number of cores is \code{detectCores() + parallel}}

\item{.export}{character vector of variables to export. Default is NULL.}

\item{.packages}{character vector of packages that tasks depends on.}

\item{progress}{The name of the file that is written into progress information. Default FALSE.}
}
\value{
A list of result.
}
\description{
Rolling apply a function on a data frame or matrix.
}
\examples{
res <- roll_apply(matrix(seq(10^4 * 3), ncol = 3), 100, function(m) eigen(cov(m))$values, parallel = -1, progress = "progress.txt")
}

