% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-all-included.R
\name{sample_all}
\alias{sample_all}
\title{A wrapper of sample}
\usage{
sample_all(x, size, replace = TRUE, prob = NULL, all = TRUE)
}
\arguments{
\item{x}{Either a vector of one or more elements from which to choose, or a positive integer}

\item{size}{a non-negative integer giving the number of  items to choose}

\item{replace}{Should sampling be with replacement?}

\item{prob}{A vector of probability for obtaining the elements of the vector being sampled}

\item{all}{If \code{replace = TRUE} and \code{all = TRUE}, sampling \code{length(x)} of items without replacement first, then sampling with replacement}
}
\value{
For sample a vector of length size with elements drawn from either x or from the integers 1:x.
}
\description{
If \code{replace = TRUE} and \code{all = TRUE}, sampling \code{length(x)} of items without replacement first,
then sampling with replacement
}
\examples{
sample_all(1:5, 6)
}

