% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{save_all}
\alias{save_all}
\title{Write all objects in the global environment to a specified file}
\usage{
save_all(file = stop("'file' must be specified"))
}
\arguments{
\item{file}{The name of the file that the objects are written to.}
}
\description{
Write all objects in the global environment to a specified file
}
\examples{
a <- 1
save_all("test.RData")
}

