% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shutdown.R
\name{shutdown}
\alias{shutdown}
\title{Shutdown the computer.}
\usage{
shutdown(times)
}
\arguments{
\item{times}{The times to shutdown the computer. Unit is second.}
}
\description{
Shutdown the computer.
}
\examples{
shutdown(10)
}

