% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_chunk.R
\name{time_chunk}
\alias{time_chunk}
\title{Time the knitr chunk.}
\usage{
time_chunk()
}
\value{
A hook function
}
\description{
Time the knitr chunk.
}
\examples{
knit_hooks$set(
  timeit = time_chunk()
)
}

