% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write_table}
\alias{write_table}
\title{Write data to a CSV file.}
\usage{
write_table(x)
}
\arguments{
\item{x}{the object to be written, preferably a Vector, matrix, or data frame}
}
\description{
Write data to a CSV file. The name of CSV file is the same with the name of the data.
}

