% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rollapply.R
\name{roll_apply}
\alias{roll_apply}
\title{Rolling apply a function on a data frame or matrix.}
\usage{
roll_apply(x, n, fun, ..., progress = "none", SIMPLIFY = FALSE)
}
\arguments{
\item{x}{A data frame or a matrix.}

\item{n}{The rolling windown size.}

\item{fun}{A function whose first parameter is the \code{x}.}

\item{...}{Additional parameter passed to \code{fun}}

\item{progress}{Create a progress bar("none", "txt", "tk", "win", "timer")}

\item{SIMPLIFY}{Attempt to return a data frame if \code{fun} returns a vector or a list whose elements are vectors with length 1.}
}
\value{
A data frame or list containing the rolling results. See \code{SIMPLIFY}
}
\description{
Rolling apply a function on a data frame or matrix.
}
\examples{
m <- matrix(1:10, ncol = 2)
roll_apply(m, 4, colMeans)
}

