% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rollapply.R
\name{roll_apply}
\alias{roll_apply}
\title{Title}
\usage{
roll_apply(x, n, fun, ...)
}
\arguments{
\item{x}{A data frame or a matrix.}

\item{n}{The rolling windown size.}

\item{fun}{A function whose first parameter is the \code{x}.}

\item{...}{Additional parameter passed to \code{fun}}
}
\value{
A list containing the rolling results
}
\description{
Title
}
\examples{
m <- matrix(1:10, ncol = 2)
roll_apply(m, 4, colMeans)
}

