% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/roll_apply_par.R
\name{roll_apply_par}
\alias{roll_apply_par}
\title{Rolling apply a function on a data frame or matrix.}
\usage{
roll_apply_par(data, width, FUN, ..., parallel = FALSE, .export = NULL,
  .packages = NULL, progress = FALSE)
}
\arguments{
\item{data}{A data frame or a matrix.}

\item{width}{The rolling windown size.}

\item{FUN}{function whose first parameter is the \code{data}.}

\item{...}{Additional parameter passed to \code{FUN}}

\item{parallel}{parallel computing. Default is FALSE. If parallel is positive,
parallel is the number of cores. If parallel is negative, the number of cores is \code{detectCores() + parallel}}

\item{.export}{character vector of variables to export. Default is NULL.}

\item{.packages}{character vector of packages that tasks depends on.}

\item{progress}{Create a progress bar. Default FALSE.}
}
\value{
A list of result.
}
\description{
Rolling apply a function on a data frame or matrix.
}
\examples{
res <- roll_apply_par(matrix(seq(10^4 * 3), ncol = 3), 100, function(m) eigen(cov(m))$values, parallel = TRUE, progress = TRUE)
}

